ledger: move catchpoint related methods into a storage package (#4835)
