+++ ledger/msgp_gen_test.go
@@ -194,66 +194,6 @@ func BenchmarkUnmarshalcatchpointFileChunkV6(b *testing.B) {
 	}
 }
 
-func TestMarshalUnmarshalcatchpointFirstStageInfo(t *testing.T) {
-	partitiontest.PartitionTest(t)
-	v := catchpointFirstStageInfo{}
-	bts := v.MarshalMsg(nil)
-	left, err := v.UnmarshalMsg(bts)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(left) > 0 {
-		t.Errorf("%d bytes left over after UnmarshalMsg(): %q", len(left), left)
-	}
-
-	left, err = msgp.Skip(bts)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(left) > 0 {
-		t.Errorf("%d bytes left over after Skip(): %q", len(left), left)
-	}
-}
-
-func TestRandomizedEncodingcatchpointFirstStageInfo(t *testing.T) {
-	protocol.RunEncodingTest(t, &catchpointFirstStageInfo{})
-}
-
-func BenchmarkMarshalMsgcatchpointFirstStageInfo(b *testing.B) {
-	v := catchpointFirstStageInfo{}
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		v.MarshalMsg(nil)
-	}
-}
-
-func BenchmarkAppendMsgcatchpointFirstStageInfo(b *testing.B) {
-	v := catchpointFirstStageInfo{}
-	bts := make([]byte, 0, v.Msgsize())
-	bts = v.MarshalMsg(bts[0:0])
-	b.SetBytes(int64(len(bts)))
-	b.ReportAllocs()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		bts = v.MarshalMsg(bts[0:0])
-	}
-}
-
-func BenchmarkUnmarshalcatchpointFirstStageInfo(b *testing.B) {
-	v := catchpointFirstStageInfo{}
-	bts := v.MarshalMsg(nil)
-	b.ReportAllocs()
-	b.SetBytes(int64(len(bts)))
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		_, err := v.UnmarshalMsg(bts)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
 func TestMarshalUnmarshalencodedBalanceRecordV5(t *testing.T) {
 	partitiontest.PartitionTest(t)
 	v := encodedBalanceRecordV5{}
